+++ local-s3-spring-boot-starter/src/main/java/com/robothy/s3/spring/boot/LocalS3AutoConfiguration.java
@@ -98,6 +98,7 @@ public class LocalS3AutoConfiguration {
     map.from(properties::getDataPath).whenHasText().to(builder::dataPath);
     map.from(properties::getBuckets).to(buckets -> builder.buckets(buckets.toArray(String[]::new)));
     map.from(properties::isInitialDataCacheEnabled).to(builder::initialDataCacheEnabled);
+    map.from(properties.getInMemory()::getMaxSize).as(DataSize::toBytes).to(builder::maxInMemoryBytes);
     map.from(properties::isCompositeMultipartEtags).to(builder::compositeMultipartEtags);
     map.from(properties::getVirtualHostDomains)
         .to(domains -> builder.virtualHostDomains(domains.toArray(String[]::new)));